error handling was fixed in farm's and pet's services

+++ Infrastructure.Services/Services/FarmsService.cs
@@ -6,8 +6,10 @@ using Application.Services.Services;
 using AutoMapper;
 using Domain.Core.Models;
 using Domain.Interfaces;
+using Infrastructure.Services.Exceptions;
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
 
 namespace Infrastructure.Services.Services
@@ -34,9 +36,13 @@ namespace Infrastructure.Services.Services
             return farm.Id;
         }
 
-        public async Task DeleteFarmByIdAsync(Guid farmId)
+        public async Task DeleteFarmByIdAsync(Guid farmId, Guid userId)
         {
             var farm = await _repositoryManager.Farms.GetFarmByIdAsync(farmId, false);
+            if (farm is null)
+                throw new EntityNotFoundException("farm was't found");
+            if (farm.UserId != userId)
+                throw new AccessException("you can't delete this farm");
             _repositoryManager.Farms.DeleteFarm(farm);
             await _repositoryManager.SaveChangeAsync();
         }
@@ -44,6 +50,8 @@ namespace Infrastructure.Services.Services
         public async Task DeleteFarmByUserIdAsync(Guid userId)
         {
             var farm = await _repositoryManager.Farms.GetFarmByUserIdAsync(userId, false);
+            if (farm is null)
+                throw new EntityNotFoundException("farm was't found");
             _repositoryManager.Farms.DeleteFarm(farm);
             await _repositoryManager.SaveChangeAsync();
         }
@@ -51,6 +59,8 @@ namespace Infrastructure.Services.Services
         public async Task<FarmReadingDto> GetFarmByUserIdAsync(Guid userId)
         {
             var farm = await _repositoryManager.Farms.GetFarmByUserIdAsync(userId, false);
+            if (farm is null)
+                throw new EntityNotFoundException("farm was't found");
             var farmDto = _mapper.Map<FarmReadingDto>(farm);
             return farmDto;
         }
@@ -58,20 +68,30 @@ namespace Infrastructure.Services.Services
         public async Task<FarmMinReadingDto> GetMinFarmByUserIdAsync(Guid userId)
         {
             var farm = await _repositoryManager.Farms.GetFarmByUserIdAsync(userId, false);
+            if (farm is null)
+                throw new EntityNotFoundException("farm was't found");
             var farmDto = _mapper.Map<FarmMinReadingDto>(farm);
             return farmDto;
         }
 
-        public async Task<FarmReadingDto> GetFarmByIdAsync(Guid farmId)
+        public async Task<FarmReadingDto> GetFarmByIdAsync(Guid farmId, Guid userId)
         {
             var farm = await _repositoryManager.Farms.GetFarmByIdAsync(farmId, false);
+            if (farm is null)
+                throw new EntityNotFoundException("farm was't found");
+            if (!IsUserFriendOfFarm(farm, userId) && farm.UserId != userId)
+                throw new AccessException("you can't get this farm");
             var farmDto = _mapper.Map<FarmReadingDto>(farm);
             return farmDto;
         }
 
-        public async Task<FarmMinReadingDto> GetMinFarmByIdAsync(Guid farmId)
+        public async Task<FarmMinReadingDto> GetMinFarmByIdAsync(Guid farmId, Guid userId)
         {
             var farm = await _repositoryManager.Farms.GetFarmByIdAsync(farmId, false);
+            if (farm is null)
+                throw new EntityNotFoundException("farm was't found");
+            if (!IsUserFriendOfFarm(farm, userId) && farm.UserId != userId)
+                throw new AccessException("you can't get this farm");
             var farmDto = _mapper.Map<FarmMinReadingDto>(farm);
             return farmDto;
         }
@@ -79,6 +99,13 @@ namespace Infrastructure.Services.Services
         public async Task<IEnumerable<FarmMinReadingDto>> GetFriendFarmsAsync(Guid userId)
         {
             var farms = await _repositoryManager.Farms.GetFriendFarmsAsync(userId);
+            var farmsDto = _mapper.Map<IEnumerable<FarmMinReadingDto>>(farms);
+            return farmsDto;
+        }
+
+        public async Task<IEnumerable<FarmMinReadingDto>> GetFriendFarmAsync(Guid userId, Guid friendId)
+        {
+            var farms = await _repositoryManager.Farms.GetFriendFarmAsync(userId, friendId);
             var farmsDto = _mapper.Map<IEnumerable<FarmMinReadingDto>>(farms);
             return farmsDto;
         }
@@ -93,11 +120,23 @@ namespace Infrastructure.Services.Services
         public async Task UpdateFarmAsync(FarmUpdatingDto farmDto, Guid userId)
         {
             var farm = await _repositoryManager.Farms.GetFarmByIdAsync(farmDto.Id, false);
+            if (farm is null)
+                throw new EntityNotFoundException("farm was't found");
             if (farm.UserId != userId)
-                throw new InvalidOperationException("you can't update this farm");
+                throw new AccessException("you can't update this farm");
             var farmForUpdating = _mapper.Map<Farm>(farmDto);
-            _repositoryManager.Farms.UpdateFarm(farm);
+            _repositoryManager.Farms.UpdateFarm(farmForUpdating);
             await _repositoryManager.SaveChangeAsync();
+        }
+
+        private bool IsUserFriendOfFarm(Farm farm, Guid userId)
+        {
+            if (farm.FarmFriends is null)
+                return false;
+            var containsUser = farm.FarmFriends
+                .Where(e => e.UserId.Equals(userId))
+                .Any();
+            return containsUser;
         }
     }
 }
